Create 7 DIGITS - 1
In this project s’ purpose is to implement a randomized counter in Assembly. A 4 digit ssd should be conneted. When my code is not counting it should be show my ID’s last digitsWhen the counter reaches 0, the number 0000 should be displayed for a second, then the code should go back to idle state waiting for the next button press. Pressing the button while counting down should pause counting, and pressing again should resume counting.