fully operational scientific calculator with C.
In this project, create a fully operational scientific calculator with C. The calculator have keypad cause to enter number and execute mathematical functions, to show display numbers and operators, the SSD is used. When on power up SSD should show ID (first 2 and last 2 digits). When press the buttons for number the SSD will be cleared , when keys are entered, the SSD will shift the numbers left, while not displaying anything for empty digits. When digits are full, new number should be ignored.